Simplification
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/turbine/trunk@8666 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/tcl/turbine/tcl-turbine.c
@@ -657,8 +657,8 @@ Turbine_Worker_Loop_Cmd(ClientData cdata, Tcl_Interp* interp,
 
   // Maintain separate buffer from xfer, since xfer may be
   // used in code that we call.
-  int buffer_size = ADLB_DATA_MAX;
-  void* buffer = malloc((size_t)buffer_size);
+  size_t buffer_size = ADLB_DATA_MAX;
+  void* buffer = malloc(buffer_size);
 
   turbine_code code = turbine_worker_loop(interp, buffer, buffer_size,
                                           work_type);
